code fixes and separation
